// src/api/landmarks/LandmarkModel.ts
export class LandmarkModel {
    id?: number;
    placeId?: string;
    name?: string;
    landmarkName?: string;
    coordinates?: string;
    completeTime?: string;
    highLight?: string;
    height?: string;
    description?: string;
    floor?: string;
    photos?: string;
    official?: string;
}